+++ static/js/server_events_dispatch.js
@@ -342,10 +342,10 @@ export function dispatch_normal_event(event) {
             composebox_typeahead.update_emoji_data();
             break;
 
-        case "realm_filters":
-            page_params.realm_filters = event.realm_filters;
-            markdown.update_linkifier_rules(page_params.realm_filters);
-            settings_linkifiers.populate_linkifiers(page_params.realm_filters);
+        case "realm_linkifiers":
+            page_params.realm_linkifiers = event.realm_linkifiers;
+            markdown.update_linkifier_rules(page_params.realm_linkifiers);
+            settings_linkifiers.populate_linkifiers(page_params.realm_linkifiers);
             break;
 
         case "realm_domains":