+++ zerver/tests/test_home.py
@@ -167,6 +167,7 @@ class HomeTest(ZulipTestCase):
         "realm_invite_to_realm_policy",
         "realm_invite_to_stream_policy",
         "realm_is_zephyr_mirror_realm",
+        "realm_linkifiers",
         "realm_logo_source",
         "realm_logo_url",
         "realm_mandatory_topics",